Renamed tests to expected names for autodetection, updated setup.py
