+++ pygoslin/tests/Shorthand_test.py
